requirements.txt
# Django Core
Django==5.1
djangorestframework==3.15.2
django-cors-headers==4.4.0
django-filter==24.3

# Database - Updated for Python 3.13 compatibility
psycopg2-binary>=2.9.10

# Authentication & Security
PyJWT==2.9.0
cryptography==42.0.8
python-decouple==3.8
pyotp==2.9.0
qrcode==8.2
Pillow>=11.3.0

# Caching
redis==5.0.8
django-redis==5.4.0

# API Documentation
drf-spectacular==0.27.2

# Task Queue (for async processing)
celery==5.4.0
kombu==5.4.0

# WebSocket Support (optional - install if needed)
# channels==4.1.0
# channels-redis==4.2.0
# daphne==4.1.2

# Production Server
gunicorn==22.0.0
whitenoise==6.7.0

# Development & Testing
pytest==8.3.2
pytest-django==4.9.0
factory-boy==3.3.1
faker==28.4.1

# Utilities
python-dateutil==2.9.0
pytz==2024.2